requirements.txt
boto3
pinecone
python-dotenv
tqdm
openai